combined phrase preprocess scripts
