+++ settings.py
@@ -1,5 +1,12 @@
-
-ASL_TRAIN_PATH = "../training_data"
+# letter training data
+ASL_TRAIN_LETTERS_PATH = "../training_data"
+# phrase training data
+ASL_TRAIN_VIDEO_PATH = "../videos"
+# image frames of phrase training data
+ASL_FRAME_PATH = "../video_output"
+# hand3d training data
 HAND_TRAIN_PATH = "../RHD_published_v2/"
+
+# letter machine learning model
 JSON_PATH = "./pickle/model.json"
 WEIGHTS_PATH = "./pickle/model.h5"
